pyproject.toml
[tool.poetry]
name = "nicoinfo"
version = "1.0.1"
description = "n站饼图推送"
authors = ["ECUsam <1254281150@qq.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "3.8.4"
aiosignal = "1.3.1"
anyio = "3.6.2"
async-timeout = "4.0.2"
attrs = "23.1.0"
beautifulsoup4 = "4.12.2"
certifi = "2023.5.7"
charset-normalizer = "3.1.0"
click = "8.1.3"
colorama = "0.4.6"
decorator = "5.1.1"
fastapi = "0.95.1"
frozenlist = "1.3.3"
h11 = "0.14.0"
httptools = "0.5.0"
idna = "3.4"
loguru = "0.6.0"
lxml = "4.9.2"
msgpack = "1.0.5"
multidict = "6.0.4"
nonebot-adapter-onebot = "2.2.3"
nonebot2 = "2.0.0rc4"
py = "1.11.0"
pydantic = "1.10.7"
pygtrie = "2.5.0"
python-dotenv = "1.0.0"
PyYAML = "6.0"
requests = "2.30.0"
retry = "0.9.2"
sniffio = "1.3.0"
soupsieve = "2.4.1"
starlette = "0.26.1"
tomli = "2.0.1"
typing-extensions = "4.5.0"
urllib3 = "2.0.2"
uvicorn = "0.22.0"
watchfiles = "0.19.0"
websockets = "11.0.3"
win32-setctime = "1.1.0"
yarl = "1.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "nicoinfo"
version = "1.0.0"
description = "n站饼图推送"
readme = "README.md"
requires-python = ">=3.8, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["nicoinfo/plugins"]
builtin_plugins = ["echo", "single_session"]

